test/integ.repo-patrol.js.snapshot/asset.68fbb7092bfabf4a2b3e4cf007227cd32a57d66c1bd024cf5e0ae91715325389/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-patrol-agent"
version = "0.1.0"
description = "AI agent for automated OSS repository maintenance"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "bedrock-agentcore>=0.1.0",
    "boto3>=1.35.0",
    "PyGithub>=2.5.0",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "moto[dynamodb,s3]>=5.0",
    "ruff>=0.8.0",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
